models/sam/pyproject.toml
[project]
name = "sam"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.14",
    "strawberry-graphql[fastapi]>=0.275.5",
    "ultralytics>=8.3.162",
]
